requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
appdirs==1.4.4
asgiref==3.8.1
attr==0.3.1
attrs==24.2.0
azure-core==1.31.0
azure-storage-blob==12.23.1
bleach==5.0.1
blinker==1.8.2
boto==2.49.0
boto3==1.35.29
botocore==1.35.29
boxing==0.1.4
branca==0.8.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
Django==4.2.16
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==2.4.0
django-migration-linter==5.1.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
dnspython==2.7.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
expiringdict==1.2.2
folium==0.18.0
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-appengine-logging==1.4.5
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.2
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.2
humansignal-drf-yasg==1.21.9
idna==3.10
ijson==3.3.0
importlib_metadata==8.4.0
inflection==0.5.1
isodate==0.6.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
label-studio==1.13.1
label-studio-sdk==1.0.7
launchdarkly-server-sdk==8.2.1
lockfile==0.12.2
lxml==5.3.0
lxml_html_clean==0.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
narwhals==1.9.4
nltk==3.9.1
numpy==1.26.4
openai==1.50.2
opencv-python==4.10.0.84
opentelemetry-api==1.27.0
ordered-set==4.0.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.10
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.11.2
pyparsing==3.1.4
pyRFC3339==1.1
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0.2
redis==3.5.3
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-mock==1.12.1
rich==13.9.2
rpds-py==0.20.1
rq==1.10.1
rsa==4.9
rules==3.4
s3transfer==0.10.2
semver==3.0.2
sentry-sdk==2.18.0
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.5.1
streamlit==1.39.0
streamlit-cropper==0.2.2
streamlit_folium==0.23.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==0.18.0
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
watchdog==5.0.3
webencodings==0.5.1
wheel==0.40.0
wrapt==1.16.0
xmljson==0.2.1
xyzservices==2024.9.0
zipp==3.20.2

#teste